backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
transformers
sentence-transformers
torch
weaviate-client
nltk
python-multipart